Latest updated code. Fixed battery percentage calculations. Fixed reed switch not turning off laser diode. Changed photores and photodiode adc values based on testing

+++ SmartWaterBottleProject/FLASC v1/Batteryreader.h
@@ -10,6 +10,6 @@
 
 #include <stdint.h>
 
-uint8_t Batteryread(void);
+uint16_t Batteryread(void);
 
 #endif /* BATTERYREADER_H_ */
